+++ RELEASE-NOTES.md
@@ -5,7 +5,7 @@
 * #41 A wrapper for BQTestRuntime injectable to JUnit
 * #44 BQModuleProviderChecker: a test helper that ensures module provider 
       is available for auto-loading
-* #46 Support classpath:/ URLs in configuration loader.
+* #46 Support classpath: URLs in configuration loader.
 * #47 Configuration type for abstract resource
 * #54 @BindingAnnotation's in Bootique should be allowed on methods